[Graph.ts] updating previously ignored method, and separation of classes
